Re-organize and simplify code for parsing network bytes
Create the `SSHProtocolParser` protocol which handles the nitty-gritty of splitting out the Data into the right amount of bytes, and extrapolating Swift data types out of the bytes. This also returns the remaining data for parsing the next field(s).

This results in, for instance in this commit RawPacketParser, far simpler code. This will help lay the groundwork for the rest of the parsing code that is to come.