Add tests for FileAttributesSerializationV3
Split tests into sections for respective methods. I'll need to do this for all other tests as well in the same way.

For some reason there isn't coverage on the decoder in that if chain I modified. Perhaps it's confused by the mixture of throw and return?

Removed unused Data+Extensions. Added tests to Result+Extensions explicitly instead of foolishly relying on other tests.

I don't have coverage on `ByteBuffer+Extensions:serialize(_ value:)` yet, but I'll do that. Covering the guard requires a `String` that is over `UInt32.max` bytes of storage, which means tests run really slow while the memory is being allocated. I started with a string repeating initializer, which was way too slow. Optimized to a Data initializer of "a", which was faster, but \0 is valid ascii/utf8 so I did a zeroed Data initializer of the length which was pretty fast. I could probably allocate closer to the bare metal, but the tests are still annoying slow with this. I'm thinking of just putting an assert or precondition in the `writeSftpString` method so I don't have to worry about the missing coverage in all of the serializers. What are the chances this would happen, realistically?