Finally get both reading and writing to work properly with a new handler
I had to ditch Combine because of several reasons. 

1. NIO won't immediately stop sending your handler new items. If you request it to stop via `read(context:)`, you still might get several (or a lot) more messages in the meantime if the socket has a lot to read.
2. Combine won't accept any queued items, even in a buffer, if the sink is not established.
3. Adding a sink will start the chain of actual subscriptions which establishes demand. Without that demand, the PassthroughSubject won't even send the data to the buffer.
4. I was unable to find a way to establish demand only from a custom buffer to the subject before the sink and subscriptions were established. If I could do that, I'm pretty sure it was be incredibly hacky and not intuitive, and ultimately not worth it.
5. Due to these restrictions on Combine (requiring a full pipeline) and the fact that I have to queue a lot of data because of NIO's read mechanism, Combine is not suitable. It can not queue the data in a way that makes sense to use Combine.

Ultimately, I had to just write a custom stream object, which can perform the duty of queueing the data, attaching a collector/handler, queue the pending futures, and handle the completion. This is a combination of the subject, buffer, demand bridge, and future sink all in one.

Additionally, I had to combine the SftpDataChannelHandler and the SftpServerChannelHandler into one handler (tentatively name SftpDataChannelHandler2). This is because the `read(context:)` must have information about both the status of the currently-executing message as well as the queued messages and the data stream in order to be accurate. Splitting into two de-coupled handlers was either impossible or I did not see an intuitive solution for this. I would have liked to keep Data for the MessagePart <=> SftpMessage section, and Server for the message handling, but alas this seems to work out better. I will accept the increased complexity of the single handler.

Cleanup of past attempts and new code to follow.